backend/requirements.txt
Flask
flask-cors
flask-jwt-extended
flask-socketio
python-dotenv
flask-pymongo
eventlet
bcrypt